+++ src/ansys/aedt/core/internal/errors.py
