+++ src/ansys/aedt/core/modeler/modeler_3d.py
@@ -29,10 +29,10 @@ import os.path
 import warnings
 
 from ansys.aedt.core.application.variables import generate_validation_errors
-from ansys.aedt.core.generic.errors import GrpcApiError
 from ansys.aedt.core.generic.file_utils import generate_unique_name
 from ansys.aedt.core.generic.file_utils import open_file
 from ansys.aedt.core.generic.general_methods import pyaedt_function_handler
+from ansys.aedt.core.internal.errors import GrpcApiError
 from ansys.aedt.core.modeler.cad.primitives_3d import Primitives3D
 from ansys.aedt.core.modeler.geometry_operators import GeometryOperators
 from ansys.aedt.core.visualization.advanced.misc import nastran_to_stl
