+++ tests/unit/test_mechanical.py
@@ -25,7 +25,7 @@
 from unittest.mock import MagicMock
 from unittest.mock import patch
 
-from ansys.aedt.core.generic.errors import AEDTRuntimeError
+from ansys.aedt.core.internal.errors import AEDTRuntimeError
 from ansys.aedt.core.mechanical import Mechanical
 import pytest
